+++ frontend/src/lib/ensureArray.ts
@@ -1,0 +1,2 @@
+export default <T>(value: T | T[]): T[] =>
+  Array.isArray(value) ? value : [value]
