+++ frontend/src/pages/musics.tsx
@@ -2,11 +2,8 @@ import * as _ from 'lodash'
 import ErrorPage from 'next/error'
 import * as React from 'react'
 
-import ResultList from '@app/components/organisms/ResultList'
-import { FormValues } from '@app/components/organisms/ResultSearchForm'
-import UserProfileLayout, {
-  Tab,
-} from '@app/components/templates/UserProfileLayout'
+import MusicsPage from '@app/components/pages/MusicsPage'
+import ensureArray from '@app/lib/ensureArray'
 import {
   parseDifficultyString,
   parsePlayStyleString,
@@ -14,7 +11,6 @@ import {
 import throwSSRError from '@app/lib/throwSSRError'
 import { PageComponentType } from '@app/pages/_app'
 import { Difficulty, PlayStyle } from '@app/queries'
-import { Router } from '@app/routes'
 
 export type RequiredQuery = {
   screenName: string
@@ -39,33 +35,7 @@ export interface Props {
   page?: number
 }
 
-const compactFormValues = ({
-  title,
-  playStyle,
-  difficulties,
-  levels,
-}: FormValues): Partial<FormValues> => {
-  const newValues: Partial<FormValues> = { playStyle }
-
-  if (title) {
-    newValues.title = title
-  }
-
-  if (difficulties.length !== 0) {
-    newValues.difficulties = difficulties
-  }
-
-  if (levels.length !== 0) {
-    newValues.levels = levels
-  }
-
-  return newValues
-}
-
-const ensureArray = <T extends any>(value: T | T[]): T[] =>
-  Array.isArray(value) ? value : [value]
-
-const MusicsPage: PageComponentType<Props, Props, Query> = ({
+const PageComponent: PageComponentType<Props, Props, Query> = ({
   screenName,
   title,
   playStyle,
@@ -77,65 +47,19 @@ const MusicsPage: PageComponentType<Props, Props, Query> = ({
     return <ErrorPage statusCode={404} />
   }
 
-  const initialValues: FormValues =
-    title == null &&
-    playStyle == null &&
-    (difficulties == null || difficulties.length === 0) &&
-    (levels == null || levels.length === 0)
-      ? {
-          title: null,
-          playStyle: PlayStyle.Sp,
-          difficulties: [],
-          levels: [12],
-        }
-      : {
-          title: title || null,
-          playStyle: playStyle || PlayStyle.Sp,
-          difficulties: difficulties || [],
-          levels: levels || [],
-        }
-
-  const replaceQuery = (newQuery: any) => {
-    const currentQuery = _.omit(Router.query || {}, 'screenName')
-    const query = { ...currentQuery, ...newQuery }
-
-    // currently next-routes doesn't support array for query parameters,
-    // so we use `Router.replace` instead of `Router.replaceRoute`.
-    Router.replace(
-      {
-        pathname: '/musics',
-        query: {
-          ...query,
-          screenName,
-        },
-      },
-      {
-        pathname: location.pathname,
-        query,
-      },
-      { shallow: true },
-    )
-  }
-
   return (
-    <UserProfileLayout screenName={screenName} activeTab={Tab.Musics}>
-      <ResultList
-        initialValues={initialValues}
-        screenName={screenName}
-        onSubmit={values => {
-          const compactedFormValues = compactFormValues(values)
-          replaceQuery({ ...compactedFormValues })
-        }}
-        onPageChange={newActivePage => {
-          replaceQuery({ page: newActivePage })
-        }}
-        defaultActivePage={page}
-      />
-    </UserProfileLayout>
+    <MusicsPage
+      screenName={screenName}
+      title={title}
+      playStyle={playStyle}
+      difficulties={difficulties}
+      levels={levels}
+      page={page}
+    />
   )
 }
 
-MusicsPage.getInitialProps = ({ res, query }) => {
+PageComponent.getInitialProps = ({ res, query }) => {
   if (!query.screenName) {
     throwSSRError(res, 404)
     return {}
@@ -158,4 +82,4 @@ MusicsPage.getInitialProps = ({ res, query }) => {
   }
 }
 
-export default MusicsPage
+export default PageComponent
